making a lot of progress here - now a functional approach
