+++ tslint.json
@@ -1,33 +1,20 @@
 {
   "rules": {
-    "align": [
-      true,
-      "parameters",
-      "statements"
-    ],
+    "align": [true, "parameters", "statements"],
     "ban": false,
     "class-name": true,
-    "comment-format": [
-      true,
-      "check-space"
-    ],
+    "comment-format": [true, "check-space"],
     "curly": true,
     "object-curly-spacing": true,
     "eofline": true,
     "forin": false,
-    "indent": [
-      true,
-      "spaces"
-    ],
+    "indent": [true, "spaces"],
     "linebreak-style": [true, "LF"],
     "interface-name": false,
     "jsdoc-format": true,
     "label-position": true,
     "label-undefined": true,
-    "max-line-length": [
-      true,
-      140
-    ],
+    "max-line-length": [true, 140],
     "member-ordering": [
       true,
       "public-before-private",
@@ -37,14 +24,7 @@
     "no-any": false,
     "no-arg": true,
     "no-bitwise": true,
-    "no-console": [
-      true,
-      "debug",
-      "info",
-      "time",
-      "timeEnd",
-      "trace"
-    ],
+    "no-console": [true, "debug", "info", "time", "timeEnd", "trace"],
     "no-construct": true,
     "no-constructor-vars": false,
     "no-debugger": true,
@@ -75,21 +55,11 @@
       "check-else",
       "check-whitespace"
     ],
-    "quotemark": [
-      true,
-      "single",
-      "jsx-double"
-    ],
+    "quotemark": [true, "single", "jsx-double"],
     "radix": true,
-    "semicolon": [
-      true,
-      "always"
-    ],
-    "switch-default": true,
-    "triple-equals": [
-      true,
-      "allow-null-check"
-    ],
+    "semicolon": [true, "always"],
+    "switch-default": false,
+    "triple-equals": [true, "allow-null-check"],
     "typedef": [
       false,
       "call-signature",
@@ -98,7 +68,8 @@
       "member-variable-declaration"
     ],
     "typedef-whitespace": [
-      true, {
+      true,
+      {
         "call-signature": "nospace",
         "index-signature": "nospace",
         "parameter": "nospace",
@@ -106,15 +77,8 @@
         "variable-declaration": "nospace"
       }
     ],
-    "use-strict": [
-      false,
-      "check-module",
-      "check-function"
-    ],
-    "variable-name": [
-      true,
-      "ban-keywords"
-    ],
+    "use-strict": [false, "check-module", "check-function"],
+    "variable-name": [true, "ban-keywords"],
     "whitespace": [
       true,
       "check-branch",
@@ -124,7 +88,5 @@
       "check-type"
     ]
   },
-  "rulesDirectory": [
-    "node_modules/tslint-eslint-rules/dist/rules"
-  ]
+  "rulesDirectory": ["node_modules/tslint-eslint-rules/dist/rules"]
 }
